fix: check configurations exists

+++ packages/melos/lib/src/common/intellij_project.dart
@@ -229,11 +229,14 @@ class IntellijProject {
   }
 
   Future<void> clean() async {
-    final melosConfigurations =
-        createGlob('$pathDotIdea/runConfigurations/melos_*.xml')
-            .listFileSystem(const LocalFileSystem());
-    await for (final file in melosConfigurations) {
-      await file.delete();
+    final melosConfigurationsPath = joinAll([pathDotIdea, 'runConfigurations']);
+    if (Directory(melosConfigurationsPath).existsSync()) {
+      final melosConfigurations =
+          createGlob('$melosConfigurationsPath/melos_*.xml')
+              .listFileSystem(const LocalFileSystem());
+      await for (final file in melosConfigurations) {
+        await file.delete();
+      }
     }
   }
